:art: show pools in icequeen page
Instead of recreating components i suggest to use vaults components and adapt the pools display according the new data added to pools : "lockForSnob" (bool) and "lockRate" . This way we are able to adapt vault diplay and possible actions according to where it's shown. Also in the long term,when we havemore pools,  filtering features will already be available in each page.